src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/gc-collected-shadowroot-crash-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: div1.createShadowRoot is not a function
